fix: failing to import xkcd_wrapper if either only requests or aiohttp were installed
- xkcd_wrapper init checks for the existence of requests and aiohttp. If none is found, throw error, otherwise just import the corresponding client (Client when requests is installed, AsyncClient when aiohttp is installed)
- Added BaseClient as a base for both clients. This way each client is just dependent on either requests or aiohttp
