+++ docs/history.md
@@ -1,5 +1,8 @@
 # History
 
+### 0.2.2 (13-08-2020)
+* Fixed failing to import xkcd_wrapper if either only requests or aiohttp were installed
+
 ### 0.2.1 (11-08-2020)
 * Separate dependencies
     (you can now use the async implementation without having to install the sync dependencies and vice versa)
